pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Szymon Walczak"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "openai-whisper (>=20250625,<20250626)",
    "langchain-openai (>=1.1.10,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)",
    "pymupdf (>=1.27.1,<2.0.0)",
    "moviepy (>=2.2.1,<3.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-core (>=1.2.16,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-huggingface (>=1.2.0,<2.0.0)",
    "sentence-transformers (>=5.2.3,<6.0.0)",
    "langchain-google-genai (>=4.2.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
